[processing] Move model related tests out to own file
No other changes
